Moved Resume Uploading Logic To `resources_service.ts`
