+++ nextstep-backend/src/controllers/resume_controller.ts
@@ -2,7 +2,8 @@ import { Request, Response } from 'express';
 import { config } from '../config/config';
 import fs from 'fs';
 import path from 'path';
-import { uploadResume, scoreResume } from '../services/resume_service';
+import { uploadResume } from '../services/resources_service';
+import { scoreResume } from '../services/resume_service';
 import multer from 'multer';
 import { CustomRequest } from "types/customRequest";
 import { handleError } from "../utils/handle_error";
