+++ nextstep-backend/src/services/resume_service.ts
@@ -1,66 +1,7 @@
 import { config } from '../config/config';
-import multer from 'multer';
 import path from 'path';
-import { randomUUID } from 'crypto';
 import fs from 'fs';
 import axios from 'axios';
-
-const createResumesStorage = () => {
-    // Ensure the directory exists
-    const resumesDirectoryPath = config.resources.resumesDirectoryPath();
-    if (!fs.existsSync(resumesDirectoryPath)) {
-        fs.mkdirSync(resumesDirectoryPath, { recursive: true });
-    }
-
-    const resumesStorage = multer.diskStorage({
-        destination: (req, file, cb) => {
-            cb(null, `${resumesDirectoryPath}/`);
-        },
-        filename: (req, file, cb) => {
-            const ext = path.extname(file.originalname);
-            const id = randomUUID();
-            cb(null, id + ext);
-        }
-    });
-
-    return multer({
-        storage: resumesStorage,
-        limits: {
-            fileSize: config.resources.resumeMaxSize()
-        },
-        fileFilter: (req, file, cb) => {
-            const allowedTypes = /pdf|doc|docx/;
-            const extname = allowedTypes.test(path.extname(file.originalname).toLowerCase());
-            const mimetype = allowedTypes.test(file.mimetype);
-
-            if (extname && mimetype) {
-                return cb(null, true);
-            } else {
-                return cb(new TypeError(`Invalid file type. Only PDF, DOC, and DOCX files are allowed.`));
-            }
-        }
-    });
-};
-
-// TODO - fix the type here
-// @ts-ignore
-const uploadResume = (req): Promise<string> => {
-    return new Promise<string>((resolve, reject) => {
-        // @ts-ignore
-        createResumesStorage().single('file')(req, {}, (error) => {
-            if (error) {
-                if (error instanceof multer.MulterError || error instanceof TypeError) {
-                    return reject(error);
-                } else if (!req.file) {
-                    return reject(new TypeError('No file uploaded.'));
-                } else {
-                    return reject(new Error('Internal Server Error'));
-                }
-            }
-            resolve(req.file.filename);
-        });
-    });
-};
 
 const scoreResume = async (resumePath: string, jobDescription?: string): Promise<number> => {
     try {
@@ -79,4 +20,4 @@ const scoreResume = async (resumePath: string, jobDescription?: string): Promise
     }
 };
 
-export { uploadResume, scoreResume }; 
+export { scoreResume };
